ref(ts): Refactor `<OrganizationApiKeysList>`
This also removes reflexbox and replaces with `<PanelTable>`
